feat: refactor logo handling; add custom header logo
