+++ backend/ee/enmedd/server/workspace/store.py
@@ -12,11 +12,14 @@ from PIL import Image
 from sqlalchemy.orm import Session
 
 from ee.enmedd.server.workspace.models import AnalyticsScriptUpload
+from ee.enmedd.server.workspace.models import Workspaces
 from enmedd.configs.constants import FileOrigin
+from enmedd.configs.constants import KV_ENTERPRISE_SETTINGS_KEY
 from enmedd.configs.constants import QUALITY
 from enmedd.configs.constants import SIZE
 from enmedd.db.models import Teamspace
 from enmedd.db.models import User
+from enmedd.db.models import Workspace
 from enmedd.file_store.file_store import get_default_file_store
 from enmedd.key_value_store.factory import get_kv_store
 from enmedd.key_value_store.interface import KvKeyNotFoundError
@@ -38,6 +41,13 @@ def load_analytics_script() -> str | None:
         return None
 
 
+def store_settings(settings: Workspaces) -> None:
+    get_kv_store().store(KV_ENTERPRISE_SETTINGS_KEY, settings.model_dump())
+
+
+_CUSTOM_ANALYTICS_SECRET_KEY = os.environ.get("CUSTOM_ANALYTICS_SECRET_KEY")
+
+
 def store_analytics_script(analytics_script_upload: AnalyticsScriptUpload) -> None:
     if (
         not _CUSTOM_ANALYTICS_SECRET_KEY
@@ -49,6 +59,9 @@ def store_analytics_script(analytics_script_upload: AnalyticsScriptUpload) -> No
 
 
 _LOGO_FILENAME = "__logo__"
+_HEADERLOGO_FILENAME = "__header_logo__"
+_TEAMSPACELOGO_FILENAME = "__teamspace_logo__"
+_LOGOTYPE_FILENAME = "__logotype__"
 _PROFILE_FILENAME = "__profile__"
 
 
@@ -68,8 +81,14 @@ def guess_file_type(filename: str) -> str:
 def upload_logo(
     db_session: Session,
     file: UploadFile | str,
+    workspace_id: int,
+    is_logotype: bool = False,
 ) -> bool:
     content: IO[Any]
+
+    workspace = db_session.query(Workspace).filter_by(id=workspace_id).first()
+    if not workspace:
+        raise HTTPException(status_code=404, detail="Workspace not found")
 
     if isinstance(file, str):
         logger.info(f"Uploading logo from local path {file}")
@@ -111,14 +130,90 @@ def upload_logo(
             logger.error(f"Error processing image: {e}")
             raise HTTPException(status_code=400, detail="Invalid image file.")
 
+    file_name = f"{workspace_id}{_LOGOTYPE_FILENAME if is_logotype else _LOGO_FILENAME}"
+
     file_store = get_default_file_store(db_session)
     file_store.save_file(
-        file_name=_LOGO_FILENAME,
+        file_name=file_name,
         content=content,
         display_name=display_name,
         file_origin=FileOrigin.OTHER,
         file_type=file_type,
     )
+
+    workspace.custom_logo = file_name
+    db_session.merge(workspace)
+    db_session.commit()
+
+    return True
+
+
+def upload_header_logo(
+    db_session: Session,
+    file: UploadFile | str,
+    workspace_id: int,
+) -> bool:
+    content: IO[Any]
+
+    workspace = db_session.query(Workspace).filter_by(id=workspace_id).first()
+    if not workspace:
+        raise HTTPException(status_code=404, detail="Workspace not found")
+
+    if isinstance(file, str):
+        logger.info(f"Uploading header logo from local path {file}")
+        if not os.path.isfile(file):
+            logger.error("File does not exist")
+            return False
+
+        with open(file, "rb") as file_handle:
+            file_content = file_handle.read()
+        content = BytesIO(file_content)
+        display_name = file
+        file_type = guess_file_type(file)
+
+    else:
+        logger.info("Uploading header logo from uploaded file")
+        if not file.filename:
+            raise HTTPException(status_code=400, detail="No file uploaded.")
+
+        try:
+            image = Image.open(file.file)
+            image.verify()
+
+            file.file.seek(0)
+            image = Image.open(file.file)
+
+            max_size = SIZE
+            image.thumbnail(max_size)
+
+            img_byte_arr = io.BytesIO()
+            image_format = image.format or "JPEG"
+            image.save(img_byte_arr, format=image_format, quality=QUALITY)
+            img_byte_arr.seek(0)
+            content = img_byte_arr
+
+            display_name = file.filename
+            file_type = file.content_type or f"image/{image_format.lower()}"
+
+        except Exception as e:
+            logger.error(f"Error processing image: {e}")
+            raise HTTPException(status_code=400, detail="Invalid image file.")
+
+    file_name = f"{workspace_id}{_HEADERLOGO_FILENAME}"
+
+    file_store = get_default_file_store(db_session)
+    file_store.save_file(
+        file_name=file_name,
+        content=content,
+        display_name=display_name,
+        file_origin=FileOrigin.OTHER,
+        file_type=file_type,
+    )
+
+    workspace.custom_header_logo = file_name
+    db_session.merge(workspace)
+    db_session.commit()
+
     return True
 
 
@@ -233,7 +328,7 @@ def upload_teamspace_logo(
             logger.error(f"Error processing image: {e}")
             raise HTTPException(status_code=400, detail="Invalid image file.")
 
-    file_name = f"{teamspace_id}{_LOGO_FILENAME}"
+    file_name = f"{teamspace_id}{_TEAMSPACELOGO_FILENAME}"
 
     file_store = get_default_file_store(db_session)
     file_store.save_file(
